Renamed `config_reader` to `config`
The gear now deals with load and save.

TODO:
- Update docs
